Added drone dynamics.
Now the goto command actually does something. It sets the target
point for the drone, and the drone will attempt to move there.

The dynamics are modelled somewhat willy nilly, as an oscillatory
second order system. Might want to take a closer look at the params
that go into it.