Drone can now top-touch and land-in-front-of the ground robots.
So I added some stuff to the command API (of course, everything is
still just experimentation. We'll want to update this so it matches
what our control group is doing).

I added a "visible region" circle, which does absolutely nothing
as of now. The plan is that it will come into play later, when we
start to obfuscate the world state measurements based on what the
drone can see.

I am still uncertain about the physical parameters of the iRobots.
See comment in code.